move paths filter logic to manager
