+++ tests/nativeapp/test_manager.py
@@ -1,8 +1,14 @@
 import os
+from pathlib import Path
 from textwrap import dedent
+from typing import List, Set
 from unittest import mock
 
 import pytest
+from click.exceptions import (
+    ClickException,
+    FileError,
+)
 from snowflake.cli.api.project.definition_manager import DefinitionManager
 from snowflake.cli.plugins.nativeapp.constants import (
     LOOSE_FILES_MAGIC_VERSION,
@@ -17,9 +23,13 @@ from snowflake.cli.plugins.nativeapp.exceptions import (
 from snowflake.cli.plugins.nativeapp.manager import (
     NativeAppManager,
     SnowflakeSQLExecutionError,
+    _get_relative_paths_to_sync,
     ensure_correct_owner,
 )
-from snowflake.cli.plugins.object.stage.diff import DiffResult
+from snowflake.cli.plugins.object.stage.diff import (
+    DiffResult,
+    filter_from_diff,
+)
 from snowflake.connector import ProgrammingError
 from snowflake.connector.cursor import DictCursor
 
@@ -94,7 +104,7 @@ def test_sync_deploy_root_with_stage(
     ]
     assert mock_execute.mock_calls == expected
     mock_stage_diff.assert_called_once_with(
-        native_app_manager.deploy_root, "app_pkg.app_src.stage", True, None
+        native_app_manager.deploy_root, "app_pkg.app_src.stage"
     )
     mock_local_diff_with_stage.assert_called_once_with(
         role="new_role",
@@ -716,3 +726,108 @@ def test_create_app_pkg_internal_distribution_no_special_comment(
         mock_warning.assert_called_once_with(
             "Continuing to execute `snow app run` on application package app_pkg with distribution 'internal'."
         )
+
+
+def exists_mock(path: Path):
+    if str(path) in ["/file", "/dir", "/dir/nested_file"]:
+        return True
+    else:
+        return False
+
+
+def is_dir_mock(path: Path):
+    if str(path) == "/dir":
+        return True
+    else:
+        return False
+
+
+# Mocking Path to mimic the following directory structure:
+# /file
+# /dir/nested_file
+@mock.patch(f"{NATIVEAPP_MODULE}.Path.is_dir", autospec=True)
+@mock.patch(f"{NATIVEAPP_MODULE}.Path.exists", autospec=True)
+@pytest.mark.parametrize(
+    "files_to_sync,remote_paths,expected_exception",
+    [
+        [["file", "dir/nested_file"], set(), None],
+        [["file", "file2", "dir/file3"], set(["file2", "dir/file3"]), None],
+        [["file", "file2"], set(), FileError],
+        [["dir/file3"], set(), FileError],
+        [["dir"], set(), ClickException],
+    ],
+)
+def test_get_full_file_paths_to_sync(
+    path_mock_exists,
+    path_mock_is_dir,
+    files_to_sync: List[Path],
+    remote_paths: Set[str],
+    expected_exception: Exception,
+):
+    path_mock_exists.side_effect = exists_mock
+    path_mock_is_dir.side_effect = is_dir_mock
+    if expected_exception is None:
+        result = _get_relative_paths_to_sync(files_to_sync, "/", remote_paths)
+        assert len(result) == len(files_to_sync)
+    else:
+        with pytest.raises(expected_exception):
+            _get_relative_paths_to_sync(files_to_sync, "/", remote_paths)
+
+
+def test_filter_from_diff():
+    diff = DiffResult()
+    diff.different = [
+        "different",
+        "different-2",
+        "dir/different",
+        "dir/different-2",
+    ]
+    diff.only_local = [
+        "only_local",
+        "only_local-2",
+        "dir/only_local",
+        "dir/only_local-2",
+    ]
+    diff.only_on_stage = [
+        "only_on_stage",
+        "only_on_stage-2",
+        "dir/only_on_stage",
+        "dir/only_on_stage-2",
+    ]
+
+    paths_to_keep = set(
+        [
+            "different",
+            "only-local",
+            "only-stage",
+            "dir/different",
+            "dir/only-local",
+            "dir/only-stage",
+        ]
+    )
+    diff = filter_from_diff(diff, paths_to_keep, True)
+
+    for path in diff.different:
+        assert path in paths_to_keep
+    for path in diff.only_local:
+        assert path in paths_to_keep
+    for path in diff.only_on_stage:
+        assert path in paths_to_keep
+
+
+# When prune flag is off, remote-only files are filtered out and a warning is printed
+@mock.patch(f"{NATIVEAPP_MODULE}.cc.warning")
+def test_filter_from_diff_no_prune(mock_warning):
+    diff = DiffResult()
+    diff.only_on_stage = [
+        "only-stage.txt",
+        "only-stage-2.txt",
+    ]
+
+    paths_to_keep = set(["only-stage.txt"])
+    diff = filter_from_diff(diff, paths_to_keep, False)
+
+    assert len(diff.only_on_stage) == 0
+    mock_warning.assert_called_once_with(
+        "The following files exist only on the stage:\n['only-stage.txt']\nUse the --prune flag to delete them from the stage."
+    )
